Create README.md
The SAS code is part of the code to explore the association of the veterans' population in each county and the county's median household income across the US and different states in 2019. The focus of this code is on data cleaning, data preparation, and visualization.

Data:
- I extracted the veterans' population, county, and state information from the excel file available on the VA website via the following link:
https://www.va.gov/vetdata/docs/GDX/GDX_FY19.xlsx
This Excel file is referenced as "VA- FY19 Summary of Expenditures by state and county.xlsx" in my code.

- The information regarding the median household income in counties of US in 2019 is also extracted from the US Census Bureau website via the following link:
https://www2.census.gov/programs-surveys/saipe/datasets/2019/2019-state-and-county/est19all.xls
This file is referenced as "2019 Poverty and Median Household Income Estimates.xlsx" in my code.

The attached code reads the data from 52 worksheets of the first Excel sheet and combines them with the data from the second worksheet to create the final table, which includes the veteran population and median household income for each county and state across the US. Next, two graph visualizations depict the association between the population of the veterans in each county and the median household income of the county across the US and in different states.